fix: Hack into Node's module loader to bypass require() in dash thumbs.
